Class NamedEntityMerging and PrepositionMerging.
