+++ ppp_questionparsing_grammatical/dependencyTree.py
@@ -162,80 +162,6 @@ class DependenciesTree:
             other.parent.child.remove(other)
         other.wordList = None
 
-    def mergeNamedEntityTagChildParent(self):
-        """
-            Merge all nodes n1, n2 such that:
-                * n1 is parent of n2
-                * n1 and n2 have a same namedEntityTag
-            Don't merge if the 2 words are linked by a conjonction
-        """
-        for child in self.child:
-            child.mergeNamedEntityTagChildParent()
-        if self.namedEntityTag == 'undef':
-            return
-        sameTagChild = set()
-        for child in self.child:
-            if child.namedEntityTag == self.namedEntityTag and not child.dependency.startswith('conj'):
-                sameTagChild.add(child)
-        for child in sameTagChild:
-            self.merge(child, True)
-
-    def mergeNamedEntityTagSisterBrother(self):
-        """
-            Merge all nodes n1, n2 such that:
-                * n1 and n2 have a same parent
-                * n1 and n2 have a same namedEntityTag
-                * n1 and n2 have a same dependency
-        """
-        for child in self.child:
-            child.mergeNamedEntityTagSisterBrother()
-        tagToNodes = {}
-        for child in self.child:
-            if child.namedEntityTag == 'undef' or child.dependency.startswith('conj'):
-                continue
-            try:
-                tagToNodes[child.namedEntityTag+child.dependency].add(child)
-            except KeyError:
-                tagToNodes[child.namedEntityTag+child.dependency] = set([child])
-        for sameTag in tagToNodes.values():
-            x = sameTag.pop()
-            for other in sameTag:
-                x.merge(other, True)
-
-    def mergeNamedEntityTag(self):
-        self.mergeNamedEntityTagChildParent()
-        self.mergeNamedEntityTagSisterBrother()
-
-
-    def mergePrepositionNode(self):
-        """
-            Merge x -> y into 'x y' if y is a preposition
-        """
-        for child in self.child:
-            child.mergePrepositionNode()
-            if child.getWords() in self.prepositionSet:
-                self.merge(child, True)
-
-    def mergePrepositionEdge(self):
-        """
-            Replace a -prep_x-> b by 'a x' -prep-> b if a is a verb, a -prep-> b otherwise
-            Replace a -agent-> b by 'a by' -agent-> b
-        """
-        for child in self.child:
-            child.mergePrepositionEdge()
-            if child.dependency.startswith('prep'): # prep_x or prepc_x
-                preposition = ' '.join(child.dependency.split('_')[1:]) # type of the prep (of, in, ...)
-                if self.isVerb():
-                    self.appendWord(preposition)
-                child.dependency = 'prep'
-            if child.dependency == 'agent':
-                assert self.isVerb()
-                self.append('by')
-
-    def mergePreposition(self):
-        self.mergePrepositionNode()
-        self.mergePrepositionEdge()
-
 class TreeGenerator:
     """
         A class to generate a dependency tree given the result of the Stanford parser.
