+++ tests/test_normalization.py
@@ -1,6 +1,7 @@
 import json
 
-from ppp_questionparsing_grammatical import computeTree, simplify, DependenciesTree, QuotationHandler, normalFormProduction, GrammaticalError
+from ppp_questionparsing_grammatical import computeTree, simplify, DependenciesTree,\
+    QuotationHandler, normalFormProduction, GrammaticalError, NamedEntityMerging, PrepositionMerging
 import data
 
 from unittest import TestCase
@@ -9,8 +10,8 @@ class StandardTripleTests(TestCase):
 
     def testAndnormalFormProduction(self):
         tree = computeTree(data.give_chief()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -59,8 +60,8 @@ class StandardTripleTests(TestCase):
 
     def testSuperlativenormalFormProduction(self):
         tree = computeTree(data.give_opera()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -99,8 +100,8 @@ class StandardTripleTests(TestCase):
 
     def testnormalFormProduction1(self):
         tree = computeTree(data.give_president_of_USA()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -125,8 +126,8 @@ class StandardTripleTests(TestCase):
         result=data.give_LSD_LIB()
         tree=computeTree(result['sentences'][0])
         handler.push(tree)
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -233,8 +234,8 @@ class StandardTripleTests(TestCase):
 
     def testnormalFormProduction3(self):
         tree = computeTree(data.give_obama_president_usa()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -299,8 +300,8 @@ class StandardTripleTests(TestCase):
 
     def testnormalFormProductionR8(self):
         tree = computeTree(data.mistake()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -338,8 +339,8 @@ class StandardTripleTests(TestCase):
 
     def testnormalFormProductionSuperl(self):
         tree = computeTree(data.tanzania()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -369,8 +370,8 @@ class StandardTripleTests(TestCase):
 
     def testnormalFormProductionSuperl2(self):
         tree = computeTree(data.car()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -400,14 +401,14 @@ class StandardTripleTests(TestCase):
 
     def testCop(self):
         tree = computeTree(data.black()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         self.assertRaises(GrammaticalError, lambda: simplify(tree))
 
     def testExists(self):
         tree = computeTree(data.king_england()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -430,8 +431,8 @@ class StandardTripleTests(TestCase):
 
     def testSemiQuestionWord1(self):
         tree = computeTree(data.roald()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -451,8 +452,8 @@ class StandardTripleTests(TestCase):
 
     def testSemiQuestionWord3(self):
         tree = computeTree(data.list_president2()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -472,8 +473,8 @@ class StandardTripleTests(TestCase):
 
     def testSemiQuestionWord4(self):
         tree = computeTree(data.capital1()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
@@ -493,8 +494,8 @@ class StandardTripleTests(TestCase):
 
     def testSemiQuestionWord5(self):
         tree = computeTree(data.capital2()['sentences'][0])
-        tree.mergeNamedEntityTag()
-        tree.mergePreposition()
+        NamedEntityMerging(tree).merge()
+        PrepositionMerging(tree).merge()
         qw = simplify(tree)
         result = normalFormProduction(tree, qw)
         self.assertEqual(result, {
